students/K3339/laboratory_works/Egor_Laptev/laboratory_work_3/requirements.txt
Django>=5.2.7
djangorestframework>=3.14.0
djoser>=2.2.0
mkdocs>=1.5.0
mkdocs-material>=9.0.0
